NEW ReactPAginate
Now React Paginate is inside the component Pagination

+++ client/src/components/products/cards/card/AdminModeCard.tsx
@@ -7,7 +7,7 @@ import { getProducts} from "../../../../redux/actions/products";
 import { ORDER } from "../Cards";
 import { resetFilterProducts } from "../../../../redux/actions/filterByCategory";
 import { getProductDetail } from "../../../../redux/actions/productDetail";
-
+import swal from "sweetalert";
 interface props {
   name: string;
   image: string;
@@ -24,12 +24,31 @@ const AdminModeCard = ({ name, image, price, id ,  AdmOrders , page }: props ) =
 
   function deleteHandler(e: React.MouseEvent<HTMLButtonElement>): void {
     e.preventDefault();
-    dispatch(deleteProduct(stringID));
-    // dispatch(resetPoducts())
-    dispatch(resetFilterProducts())
-    dispatch(getProducts())
-    page(1)
-    AdmOrders(stringID)
+    swal({
+      title: "Are you sure?",
+      text: "Once deleted, you will not be able to recover this product!",
+      icon: "warning",
+      dangerMode: true,
+      buttons: {
+        cancel: true,
+        confirm: true
+      }
+    }).then((value) => {
+      if (value) {
+
+        dispatch(deleteProduct(stringID));
+        dispatch(resetFilterProducts())
+        dispatch(getProducts())
+        page(1)
+        AdmOrders(stringID)
+
+        swal({
+          text: "Product deleted",
+          icon: "success"
+        })
+      }
+    })
+
   }
 
   function handleClickEdit (e: React.MouseEvent<HTMLButtonElement>) : void {
